requirements.txt
Django==3.2.8
djangorestframework==3.12.4
openai==0.10.3
Pillow==8.3.2
